Feat: Apply code review #2402 (#2383).
- Move position of `check_graph_status()` function.
- Modify the tests to not use `unittest`.
