+++ pgmpy/tests/help_functions.py
@@ -3,21 +3,3 @@ def recursive_sorted(li):
     for i in range(len(li)):
         li[i] = sorted(li[i])
     return sorted(li)
-
-
-def check_graph_status(
-    graph,
-    node_count: int,
-    edge_count: int,
-    exposures: set,
-    outcomes: set,
-    latents: set,
-    roles: dict,
-):
-    """Common graph state checking function."""
-    assert len(graph.nodes) == node_count
-    assert len(graph.edges) == edge_count
-    assert graph.exposures == exposures
-    assert graph.outcomes == outcomes
-    assert graph.latents == latents
-    assert graph.get_role_dict() == roles
